Hadoop integration fix #106
api services improved
connection pages modified
process docker container prepared
scheduler send mail fix
ui connection page modified

+++ src/api/src/application/operation/DeleteDataOperation/DeleteDataOperationCommandHandler.py
@@ -2,12 +2,15 @@ from injector import inject
 from pdip.cqrs import Dispatcher
 from pdip.cqrs import ICommandHandler
 from pdip.data.repository import RepositoryProvider
+from pdip.exceptions import OperationalException
 
 from src.application.notification.SendNotification.SendNotificationCommand import SendNotificationCommand
 from src.application.notification.SendNotification.SendNotificationRequest import NotificationAdditionalData, \
     SendNotificationRequest
 from src.application.operation.DeleteDataOperation.DeleteDataOperationCommand import DeleteDataOperationCommand
 from src.application.operation.services.DataOperationService import DataOperationService
+from src.application.schedule.DeleteCronJob.DeleteCronJobCommand import DeleteCronJobCommand
+from src.application.schedule.DeleteCronJob.DeleteCronJobRequest import DeleteCronJobRequest
 
 
 class DeleteDataOperationCommandHandler(ICommandHandler[DeleteDataOperationCommand]):
@@ -23,9 +26,18 @@ class DeleteDataOperationCommandHandler(ICommandHandler[DeleteDataOperationComma
         self.data_operation_service = data_operation_service
 
     def handle(self, command: DeleteDataOperationCommand):
+        data_operation = self.data_operation_service.get_by_id(id=command.request.Id)
+        if data_operation is  None:
+            raise OperationalException(f"{command.request.Id} Data Operation not found")
+        self.delete_exiting_cron_jobs(operation_name=data_operation.Name)
         result = self.data_operation_service.delete_data_operation(command.request.Id)
         self.repository_provider.commit()
         self.notify(message=result, id=command.request.Id)
+
+    def delete_exiting_cron_jobs(self, operation_name: str):
+        req = DeleteCronJobRequest(OperationName=operation_name)
+        command = DeleteCronJobCommand(request=req)
+        self.dispatcher.dispatch(command)
 
     def notify(self, message: str, id: int):
         data_list = []
